Support stripping 3rd language when reading input files
The boolean constant STRIP_3RD_LANGUAGE controls whether the stripping
is to be done. If yes, everything starting with the last '/' (slash)
character of the "meaning" part of the line is stripped. This means
that with this setting, the content you regularly place after a slash
won't show up. I use this to contain translations in a third language,
hence the name.